In GUI package updated HeaderView, added toString to Role enum, changed Conference constructor.
Conference constructor now has all of the deadlines listed out instead of in Map form as requested.  Role enum now has a client-friendly tostring output and an associated way to convert back from the string to the enum.  Conference DAO was changed so that a List<> instead of Collectioin<> was returned for getConferences().